+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/javadocvariable/InputJavadocVariableTagsMethods2.java
@@ -1,0 +1,110 @@
+/*
+JavadocVariable
+accessModifiers = (default)public,protected,package,private
+ignoreNamePattern = (default)null
+tokens = (default)ENUM_CONSTANT_DEF
+
+*/
+
+package com.puppycrawl.tools.checkstyle.checks.javadoc.javadocvariable;
+
+import java.io.IOException;
+
+class InputJavadocVariableTagsMethods2
+{
+    /** @param aOne Perfectly legal **/
+    void method14(int aOne)
+    {
+    }
+
+    /** @throws java.io.IOException
+     *               just to see if this is also legal **/
+    void method14()
+       throws java.io.IOException
+    {
+    }
+
+
+
+    // Test static initialiser
+    static
+    {
+        int x = 1; // should not require any javadoc
+    }
+
+    // test initialiser
+    {
+        int z = 2; // should not require any javadoc
+    }
+
+    /** handle where variable declaration over several lines **/
+    private static final int
+        ON_SECOND_LINE = 2;
+
+
+    /**
+     * Documenting different causes for the same exception
+     * in separate tags is OK (bug 540384).
+     *
+     * @throws java.io.IOException if A happens
+     * @throws java.io.IOException if B happens
+     **/
+    void method15()
+       throws java.io.IOException
+    {
+    }
+
+    /** {@inheritDoc} **/
+    public String toString()
+    {
+        return super.toString();
+    }
+
+    /** getting code coverage up **/
+    static final int serialVersionUID = 666;
+
+    //**********************************************************************/
+    // Method Name: method16
+    /**
+     * handle the case of an elaborate header surrounding javadoc comments
+     *
+     * @param aOne valid parameter content
+     */
+    //**********************************************************************/
+    void method16(int aOne)
+    {
+    }
+
+
+    /**
+     * @throws ThreadDeath although bad practice, should be silently ignored
+     * @throws ArrayStoreException another r/t subclass
+     * @throws IllegalMonitorStateException should be told to remove from throws
+     */
+    void method17()
+        throws IllegalMonitorStateException
+    {
+    }
+
+    /**
+     * declaring the imported version of an Exception and documenting
+     * the full class name is OK (bug 658805).
+     * @throws java.io.IOException if bad things happen.
+     */
+    void method18()
+        throws IOException
+    {
+        throw new IOException("to make compiler happy");
+    }
+
+    /**
+     * reverse of bug 658805.
+     * @throws IOException if bad things happen.
+     */
+    void method19()
+        throws java.io.IOException
+    {
+        throw new IOException("to make compiler happy");
+    }
+
+}
